backend/requirements.txt
# ============================================================================
# TBAPS Backend Requirements — Production
# Python 3.11+ / 3.13 compatible
# Last updated: 2026-02-18
# ============================================================================

# Core Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
python-multipart>=0.0.9

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
cryptography>=42.0.0

# OAuth2 Integrations
authlib>=1.3.0
httpx>=0.26.0
aiohttp>=3.9.0

# Background Tasks
celery[redis]>=5.3.4
redis>=5.0.0
flower>=2.0.1

# Rate Limiting
slowapi>=0.1.9

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Google APIs
google-auth>=2.26.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0

# Microsoft Graph API
msal>=1.26.0
msgraph-core>=1.0.0

# Jira
jira>=3.6.0

# Asana (5.0.4 was yanked — use latest stable)
asana>=5.0.5

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7
sentry-sdk[fastapi]>=1.39.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
tenacity>=8.2.3

# Data Science & Statistics (for bias detection)
numpy>=1.26.0
scipy>=1.11.0
tabulate>=0.9.0

# Email — NEF Agent distribution
aiosmtplib>=3.0.0
